Warning: this branch commit is largely untested, and known to throw exceptions as database structure retrieval is currently missing!
Further work on SPMySQLFramework integration:
 - Improve SPMySQL framework build settings including correct ppc builds and a Distribution configuration for the build distributions to match
 - Add new convenience querying and result methods to the framework
 - Amend Sequel Pro source to use the new SPMySQL.framework methods everywhere, replacing MCPKit methods where they differ and improving some functions
 - Remove MCPKit from the source
 - Fix a number of warnings on Release-style builds

+++ Source/ImageAndTextCell.m
@@ -53,7 +53,7 @@
 		imageFrame.size = [image size];
 		imageFrame.origin = cellFrame.origin;
 		imageFrame.origin.x += ((1 - MIN(1,INDENT_AMOUNT)) * 3) + (INDENT_AMOUNT * _indentationLevel);
-		imageFrame.origin.y += ceil((cellFrame.size.height - imageFrame.size.height) / 2);
+		imageFrame.origin.y += ceilf((cellFrame.size.height - imageFrame.size.height) / 2);
 		return imageFrame;
 	}
 	else
@@ -115,9 +115,9 @@
 		imageFrame.size = imageSize;
 
 		if ([view isFlipped])
-			imageFrame.origin.y += ceil((cellFrame.size.height + imageFrame.size.height) / 2);
+			imageFrame.origin.y += ceilf((cellFrame.size.height + imageFrame.size.height) / 2);
 		else
-			imageFrame.origin.y += ceil((cellFrame.size.height - imageFrame.size.height) / 2);
+			imageFrame.origin.y += ceilf((cellFrame.size.height - imageFrame.size.height) / 2);
 
 		imageFrame.origin.y -= 1;
 
@@ -154,9 +154,9 @@
 		imageFrame.size = imageSize;
 
 		if ([controlView isFlipped])
-			imageFrame.origin.y += ceil((cellFrame.size.height + imageFrame.size.height) / 2);
+			imageFrame.origin.y += ceilf((cellFrame.size.height + imageFrame.size.height) / 2);
 		else
-			imageFrame.origin.y += ceil((cellFrame.size.height - imageFrame.size.height) / 2);
+			imageFrame.origin.y += ceilf((cellFrame.size.height - imageFrame.size.height) / 2);
 
 		[image compositeToPoint:imageFrame.origin operation:NSCompositeSourceOver];
 	}
